Update README.md
This project integrates the MAX30102 sensor with internet connectivity using the RP2040 board. Sensor data is sent to an online database and an interactive dashboard on Thingspeak for remote monitoring and analysis. The setup includes configuring a stable Wi-Fi connection, enabling the RP2040 to communicate with other devices and access online resources. The heart monitoring system captures accurate heart rate data, using the MAX30105 library for sensor interaction and WiFiNINA for wireless communication, and sends data to Thingspeak every 15 seconds, providing real-time monitoring through the cloud.